moved static helpers 'irrelevant' for models, in test targets to a separate file
