+++ Tests/TestCases/UnitTests/RadixStack/0. Layer 7 - APPLICATION (Transactions, Balances)/NetworkEpicTests/SubmitAtomEpicTests/SubmitAtomEpicTests.swift
@@ -344,65 +344,3 @@ final class AtomStatusObservationCanceller: AtomStatusObservationCancelling {
     }
     
 }
-
-extension SignedAtom {
-    static var irrelevant: Self {
-        let unsignedAtom = UnsignedAtom.irrelevant
-        return unsignedAtom.signed(signature: .irrelevant, signatureId: .irrelevant)
-        
-    }
-}
-
-extension Signature {
-    static var irrelevant: Self {
-        try! Signature.init(r: 1, s: 1)
-    }
-}
-
-extension UnsignedAtom {
-    static var irrelevant: Self {
-        try! UnsignedAtom(atomWithPow: .irrelevant)
-    }
-}
-
-extension AtomWithFee {
-    static var irrelevant: Self {
-        try! AtomWithFee(atomWithoutPow: Atom.irrelevant, proofOfWork: .irrelevant)
-    }
-}
-
-extension Atom {
-    static var irrelevant: Self {
-        withDestinationAddress(.irrelevant)
-    }
-    
-    static func withDestinationAddress(_ address: Address, date: Date = .init()) -> Self {
-        let particle = ResourceIdentifierParticle(resourceIdentifier: .withAddress(address))
-        let particleGroup = try! ParticleGroup(spunParticles: particle.withSpin(.up))
-        
-        return Atom(
-            metaData: .timestamp(date),
-            particleGroups: [particleGroup]
-        )
-    }
-}
-
-extension ProofOfWork {
-    static var irrelevant: Self {
-        Self(seed: Data(), targetNumberOfLeadingZeros: 1, magic: 1, nonce: 0)
-    }
-}
-
-extension EUID {
-    static var irrelevant: Self {
-        try! EUID(int: 1)
-    }
-}
-
-extension ResourceIdentifier {
-    static let irrelevant: Self = .withAddress(.irrelevant)
-    
-    static func withAddress(_ address: Address) -> Self {
-        "/\(address)/\(UUID().uuidString.suffix(5))"
-    }
-}
